+++ pajbot/bot.py
@@ -137,9 +137,11 @@ class Bot:
         if self.bot_user_id is None:
             raise ValueError("The bot login name you entered under [main] does not exist on twitch.")
 
-        self.streamer_user_id = self.twitch_helix_api.get_user_id(self.streamer)
-        if self.streamer_user_id is None:
+        self.broadcaster = self.twitch_helix_api.get_user_basics_by_login(self.streamer)
+        if self.broadcaster is None:
             raise ValueError("The streamer login name you entered under [main] does not exist on twitch.")
+
+        self.streamer_user_id = self.broadcaster.id
 
         self.streamer_access_token_manager = UserAccessTokenManager(
             api=self.twitch_id_api, redis=RedisManager.get(), username=self.streamer, user_id=self.streamer_user_id
@@ -354,6 +356,14 @@ class Bot:
             return getattr(extra["command"].data, key)
         except:
             log.exception("Caught exception in get_source_value")
+
+        return None
+
+    def get_broadcaster_value(self, key, extra={}):
+        try:
+            return getattr(self.broadcaster, key)
+        except:
+            log.exception("Caught exception in get_broadcaster_value")
 
         return None